[Build] Reduced build warnings on windows

+++ src/tools/map-extractor/mpq_libmpq.h
@@ -57,38 +57,6 @@ class MPQArchive
          */
         void close();
 
-        //uint32 HashString(const char* Input, uint32 Offset)
-        //{
-        //    uint32 seed1 = 0x7fed7fed;
-        //    uint32 seed2 = 0xeeeeeeee;
-
-        //    for (uint32 i = 0; i < strlen(Input); i++)
-        //    {
-        //        uint32 val = toupper(Input[i]);
-        //        seed1 = mpq_a.buf[Offset + val] ^ (seed1 + seed2);
-        //        seed2 = val + seed1 + seed2 + (seed2 << 5) + 3;
-        //    }
-
-        //    return seed1;
-        //}
-        //mpq_hash GetHashEntry(const char* Filename)
-        //{
-        //    uint32 index = HashString(Filename, 0);
-        //    index &= mpq_a.header->hashtablesize - 1;
-        //    uint32 name1 = HashString(Filename, 0x100);
-        //    uint32 name2 = HashString(Filename, 0x200);
-
-        //    for (uint32 i = index; i < mpq_a.header->hashtablesize; ++i)
-        //    {
-        //        mpq_hash hash = mpq_a.hashtable[i];
-        //        if (hash.name1 == name1 && hash.name2 == name2) return hash;
-        //    }
-
-        //    mpq_hash nullhash;
-        //    nullhash.blockindex = 0xFFFFFFFF;
-        //    return nullhash;
-        //}
-
         /**
          * @brief
          *
@@ -101,7 +69,7 @@ class MPQArchive
             libmpq__off_t size, transferred;
             libmpq__file_unpacked_size(mpq_a, filenum, &size);
 
-            char* buffer = new char[size];
+            char* buffer = new char[(int)size];
 
             libmpq__file_read(mpq_a, filenum, (unsigned char*)buffer, size, &transferred);
 
@@ -178,13 +146,13 @@ class MPQFile
          *
          * @return size_t
          */
-        size_t getSize() { return size; }
+        size_t getSize() { return (size_t)size; }
         /**
          * @brief
          *
          * @return size_t
          */
-        size_t getPos() { return pointer; }
+		size_t getPos() { return (size_t)pointer; }
         /**
          * @brief
          *
